Create README.md
Hey Guys, in this project i will be coding an Animated Website using HTML, CSS and JavaScript. We will make use of Parallax.js and GSAP JavaScript Libraries to achieve the animations. I will not be making this project responsive because i just wanted to explore GSAP Animations So ...Maybe I'll make it Whenever I'll get time..THANKS FOR READING. :)